pyproject.toml
[tool.poetry]
name = "crawlingrobot"
version = "0.1.0"
description = ""
authors = ["biancaziliotto <b.ziliotto@studenti.unipi.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mujoco = "^3.3.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
mediapy = "^1.2.2"
torch = "^2.6.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
